+++ vico/compose/src/main/java/com/patrykandpatrick/vico/compose/cartesian/decoration/HorizontalBox.kt
@@ -29,7 +29,7 @@ import com.patrykandpatrick.vico.core.common.data.ExtraStore
 /** Creates and remembers a [HorizontalBox]. */
 @Composable
 public fun rememberHorizontalBox(
-  y: (ExtraStore) -> ClosedFloatingPointRange<Float>,
+  y: (ExtraStore) -> ClosedFloatingPointRange<Double>,
   box: ShapeComponent,
   labelComponent: TextComponent? = null,
   label: (ExtraStore) -> CharSequence = { HorizontalBox.getLabel(y(it)) },
